Total allignment
IMPORTANT COMMIT - continue from this point!!!
========================
* Created and implemented infrastructure for many necessary classes in "engine" package

* Updated prototypes of methods requred in "DB" package, made DBHandler to implement classic singleton DP

*Fixed implementations of Scenario, Event and few other classes

+++ src/engine/Action.java
@@ -4,7 +4,7 @@ import java.lang.reflect.Type;
 
 
 public class Action {
-	private int id;
+	private short id;
 	private String name;
 	private String type;
 	private Object parameter;
@@ -39,12 +39,12 @@ public class Action {
 		return this.parameter.getClass();
 	}
 	
-	public void setId(int id)
+	public void setId(short id)
 	{
 		this.id = id;
 	}
 	
-	public int getId()
+	public short getId()
 	{
 		return this.id;
 	}
